Retry failed builds / detect x86 pkgs properly
Some KDE packages get stuck with 0% CPU usage when cross compiling with
pmbootstrap (see postmarketOS/pmbootstrap#1458). There's a workaround
in pmbootstrap now, which kills processes that do not output anything
for five minutes (postmarketOS/pmbootstrap#1495). This commit changes
the repo build scripts to automatically retry the build in case of
such an error (also in case of other errors, such as download errors).

In addition, there's a bug fix that lets the repo scripts detect x86
packages properly. Previously, we ran grep $arch on the arch= line,
which means it detected x86_64 as x86. Now we have a new
get_packages.py script, that does it properly.